chores: Migrate tooltips
