+++ VocaDbWeb/New/Helpers/markdown.tsx
@@ -1,9 +1,6 @@
-import EntryToolTip from '@/components/ToolTips/EntryToolTip';
+import EntryLink from '@/components/ToolTips/EntryLink';
 import htmlParse, { Element, DOMNode, domToReact } from 'html-react-parser';
 import Link from 'next/link';
-import useSWR from 'swr';
-import { apiGet } from './FetchApiHelper';
-import { useState } from 'react';
 
 // the workaround: https://github.com/remarkablemark/html-react-parser/issues/616
 // the bug: https://github.com/remarkablemark/html-react-parser/issues/633
@@ -12,54 +9,6 @@ const isElement = (domNode: DOMNode): domNode is Element => {
 	const hasAttributes = (domNode as Element).attribs !== undefined;
 
 	return isTag && hasAttributes;
-};
-
-interface EntryLinkProps {
-	itemUrl: string;
-	children: JSX.Element | JSX.Element[] | string;
-}
-
-const EntryLink = ({ itemUrl, children }: EntryLinkProps) => {
-	const [hovered, setHovered] = useState(false);
-	const apiUrl = itemUrl
-		.replace('/Ar/', '/artists/')
-		.replace('/Al/', '/albums/')
-		.replace('/S/', '/songs/');
-	const { data, isLoading } = useSWR(
-		hovered ? `/api${apiUrl}?fields=AdditionalNames` : null,
-		apiGet,
-		{
-			revalidateOnFocus: false,
-			revalidateIfStale: false,
-		}
-	);
-	const entry = itemUrl.includes('/Ar/')
-		? 'artist'
-		: itemUrl.includes('/Al/')
-		? 'album'
-		: itemUrl.includes('/S/')
-		? 'song'
-		: 'tag';
-
-	const child = (
-		<Link
-			onMouseOver={() => {
-				if (!hovered) setHovered(true);
-			}}
-			href={itemUrl}
-		>
-			{children}
-		</Link>
-	);
-
-	return isLoading || !hovered ? (
-		child
-	) : (
-		//@ts-ignore
-		<EntryToolTip artist={data} album={data} song={data} entry={entry}>
-			{child}
-		</EntryToolTip>
-	);
 };
 
 export default function parse(html: string): string | JSX.Element | JSX.Element[] {
